requirements.txt
albumentations==1.4.2
anyio==4.3.0
astor==0.8.1
asttokens==2.4.1
attrdict==2.0.1
Babel==2.14.0
backcall==0.2.0
bce-python-sdk==0.9.6
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.1.1
cssselect==1.2.0
cssutils==2.10.2
cycler==0.12.1
Cython==3.0.10
debugpy==1.6.7
decorator==5.1.1
einops==0.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.4
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.51.0
fsspec==2023.4.0
future==1.0.0
gdown==4.4.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.0
imgaug==0.4.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.3
ipython==8.12.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.4.1
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.1
numpy==1.24.4
# nvidia-cublas-cu11==11.11.3.6
# nvidia-cuda-cupti-cu11==11.8.87
# nvidia-cuda-nvrtc-cu11==11.8.89
# nvidia-cuda-runtime-cu11==11.8.89
# nvidia-cudnn-cu11==8.7.0.84
# nvidia-cufft-cu11==10.9.0.58
# nvidia-curand-cu11==10.3.0.86
# nvidia-cusolver-cu11==11.4.1.48
# nvidia-cusparse-cu11==11.7.5.86
# nvidia-nccl-cu11==2.19.3
# nvidia-nvtx-cu11==11.8.86
opencv-contrib-python
opencv-python
opencv-python-headless
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==24.0
paddleocr
paddlepaddle
pandas
parso==0.8.4
pdf2docx==0.5.8
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==4.2.0
prefetch-generator==1.0.1
premailer==3.10.0
prompt-toolkit==3.0.42
# protobuf==5.26.1
protobuf==3.20.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyclipper==1.3.0.post5
pycryptodome==3.20.0
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0
python-docx==1.1.0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.8.1
rarfile==4.2
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
setuptools==68.2.2
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.2
sympy==1.12
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2023.7.10
torch
torchaudio
torchvision
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
# triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vietocr
visualdl==2.5.3
wcwidth==0.2.13
Werkzeug==3.0.2
wheel==0.41.2
zipp==3.17.0
python-multipart
reportlab
fastapi
uvicorn